+++ src/main/java/com/elmakers/mine/bukkit/traders/FloatItemAttr.java
@@ -1,7 +1,0 @@
-package com.elmakers.mine.bukkit.traders;
-
-public class FloatItemAttr extends NBTItemAttr {
-	public FloatItemAttr(String key, String tag, String subtag) {
-		super(key, tag, subtag, "0", "0.0");
-	}
-}