+++ src/main/java/com/elmakers/mine/bukkit/traders/IntegerItemAttr.java
@@ -1,7 +1,0 @@
-package com.elmakers.mine.bukkit.traders;
-
-public class IntegerItemAttr extends NBTItemAttr {
-	public IntegerItemAttr(String key, String tag, String subtag) {
-		super(key, tag, subtag, "0");
-	}
-}